requirements_py313.txt
# Core Dependencies - Python 3.13 compatible versions
ib_insync==0.9.86
aiohttp==3.11.11
xgboost==2.1.3
pandas==2.2.3
numpy==2.2.1
scipy==1.14.1
psycopg2-binary==2.9.10
redis==5.2.1
discord.py==2.4.0
prometheus-client==0.21.1
pyyaml==6.0.2
python-dotenv==1.0.1
joblib==1.4.2
scikit-learn==1.6.1
pytest==8.3.4
pytest-asyncio==0.25.2
black==24.12.1  